feat(FileManager): csv 파일을 읽어와라
- readFileLines() 메서드의 프로덕션 코드를 작성하였다.
